requirements.txt
astroid==2.2.5
autopep8==1.4.4
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Script==2.0.6
Flask-WTF==0.14.2
gunicorn==19.9.0
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pylint==2.3.1
six==1.12.0
typed-ast==1.3.5
visitor==0.1.3
Werkzeug==0.15.4
wrapt==1.11.1
WTForms==2.2.1
